+++ issuetracker/CHANGES.rst
@@ -2,6 +2,8 @@
 ====================
 
 - Issue information is now cached
+- Custom issue trackers must now connect to the ``issuetracker-resolve-issue``
+  event, the builtin ``missing-reference`` event is no longer used.
 
 
 0.6 (Jan 04, 2011)